src/requirements.txt
appdirs==1.4.4
attrs==19.3.0
backcall==0.2.0
black==19.10b0
certifi==2020.6.20
cffi==1.14.2
click==7.1.2
cryptography==3.0
decorator==4.4.2
envparse==0.2.0
eradicate==1.0
flake8==3.8.3
flake8-black==0.2.1
flake8-bugbear==20.1.4
flake8-eradicate==0.4.0
flake8-isort==4.0.0
flake8-print==3.1.4
freezegun==0.3.15
iniconfig==1.0.1
ipython==7.17.0
ipython-genutils==0.2.0
isort==5.4.2
jedi==0.17.2
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
parso==0.7.1
pathspec==0.8.0
peewee==3.13.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.6
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
pytest==6.0.1
pytest-env==0.6.2
pytest-freezegun==0.4.2
pytest-mock==3.3.0
python-dateutil==2.8.1
python-telegram-bot==12.8
regex==2020.7.14
six==1.15.0
testfixtures==6.14.1
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.2.5
